requirements.txt
albumentations==1.3.1
efficientnet-pytorch==0.7.1
einops==0.6.1
elasticdeform==0.5.0
eyepie==0.3.7
joblib==1.3.1
matplotlib==3.7.1
monai==1.2.0
opencv-python==4.8.0.74
Pillow==10.0.0
pretrainedmodels==0.7.4
PyYAML==6.0
relaynet-pytorch==1.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
segmentation-models-pytorch==0.3.3
tensorboard==2.13.0
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
